requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.9.0
google-generativeai==0.7.2